A class to connstruct, execute and analyse a genetic algorithm.

This class will store data during the execution of a genetic algorithm. As well as this, it contains a number of methods to interact with the algorithm and UI functionality to view the results and ammend parameters during execution.

Details

To get started, using the new() method to begin.

Public fields

par

a named list containing function parameters

dim

the dimensions of the population array

population

the population array

operators

an ordered list of the operator functions

dependents

function used by the operators

metrics

a list containing custom metrics stored during the execution of the simulation

output_graphs

a list of functions that produce a 2D graph (UI only)

time

the number of operators that have been executed

paused

whether the simulation is paused or not (UI only)

speed

the number of generations executed every second (UI only)

store_data

boolean which determines the amount of information which is stored

stored_df

data frame used to store metric data

current_generation

the number of times that every operators has been executed

shiny_ui

the UI code (UI only)

shiny_server

variable to store the shiny server function

Methods

Public methods

ga$new()

ga$add_operators()

ga$add_operator()

ga$add_metric()

ga$add_dependents()

ga$add_dependent()

ga$next_operator()

ga$next_iteration()

ga$generate_server()

ga$add_output_graph()

ga$add_output_graphs()

ga$run_ui()

ga$save_data()

ga$clone()

Method new()

This method allows you to initialize a ga class instance.

Usage

ga$new(dim, initial, parameters = list(), store_data = FALSE)

Arguments

dim

the dimensions of the population array, with the last dimension being the size of the population.

initial

the function used to generate the initial population.

parameters

a named list containing parameters used in the operator and dependancy function

store_data

a boolean value which indicates how much data is stored during the simulation.

Examples

brach <- gapackage::ga$new(
dim = c(10, 1000),
parameters = list(
 initial_min     = 0,
 B_x             = 180,
 initial_max     = 40,
 A_y             = 40,
 geno_length     = 2,
 g               = 9.81,
 remove          = 0.5,
 add_proportion  = 0.5,
 mutation_size   = 0.0000007,
 generations     = 100,
 population_size = 3000,
 maximise        = FALSE,
 location        = 0,
 minkowski_p     = 100
 ),
 store_data = FALSE,
 initial = initial_rand_uni_3d
);

Method add_operators()

A method to add an operator / operators to the genetic algorithm.

Usage

ga$add_operators(operators)

Arguments

operators

a single function or a list of functions which have an operator structure.

Examples

brach$add_operators(
  operators = list(
    mutation_cauchy,
    crossover_basic_3d
  )
)

brach$add_operators(
   mutation_cauchy
)

Method add_operator()

A method to add single operator to the genetic algorithm

Usage

ga$add_operator(operator)

Arguments

operator

a single function which has an operator structure.

Details

It is recommended that you use the add_operators() function instead.

Method add_metric()

A method to add a metric to track during the simulation

Usage

ga$add_metric(name, value)

Arguments

name

the name of the metric you want to track.

value

the expression corresponding to the value of the metric.

Details

This should only be run in either a dependency or an operator function.

Examples

self$add_metric(name = "average_fitness", value = sum(fitnesses) / length(fitnesses));

Method add_dependents()

A method to add a single dependency or a list of multiple dependencies

Usage

ga$add_dependents(dependents)

Arguments

dependents

a single dependency or a list of multiple dependancies

Examples

brach$add_dependents(
  dependents = list(
    fitness_function_single = utility_fitness_brachistochrone_3d,
    fitness_function = utility_fitness_population_3d
  )
)

Method add_dependent()

A method to add a single dependency function.

Usage

ga$add_dependent(dependent)

Arguments

dependent

a named list containing one dependent.

Examples

brach$add_dependents(
  dependents = list(
    fitness_function_single = utility_fitness_brachistochrone_3d
  )
)

Method next_operator()

A method to run the next operator in the ordered operators list.

Usage

ga$next_operator()

Examples

brach$next_operator()

Method next_iteration()

A method to run through all the operators once.

Usage

ga$next_iteration()

Examples

brach$next_iteration()

Method generate_server()

A method to generate the R Shiny server function (UI Only)

Usage

ga$generate_server()

Returns

a function intended to be used as an R Shiny server function.

Examples

brach$generate_server()

Method add_output_graph()

a method to add an output graph to the output tab in the UI (UI Only)

Usage

ga$add_output_graph(parameters)

Arguments

parameters

a named list containing a function and chart output settings

Examples

brach$add_output_graph(
 list(
   xlab = "horizontal distance",
   ylab = "vertical distance",
   FUN = manhattan_plot
 )
)

Method add_output_graphs()

a method to add multiple output graphs to the output tab in the UI (UI Only)

Usage

ga$add_output_graphs(parameters)

Arguments

parameters

a list containing many named lists which contain multiple function and chart output settings

Examples

brach$add_output_graphs(
 list(
   list(
     xlab = "horizontal distance",
     ylab = "vertical distance",
     FUN = manhattan_plot
   ),
   list(
     xlab = "horizontal distance",
     ylab = "vertical distance",
     FUN = manhattan_plot
   )
 )
)

Method run_ui()

A method to open the UI.

Usage

ga$run_ui()

Details

You should only have one localhost open at any particular time.

Method save_data()

A method to save the data stored during the simulation

Usage

ga$save_data(path)

Arguments

path

the path of the csv file that will be created.

Method clone()

The objects of this class are cloneable with this method.

Usage

ga$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `ga$new`
## ------------------------------------------------

brach <- gapackage::ga$new(
dim = c(10, 1000),
parameters = list(
 initial_min     = 0,
 B_x             = 180,
 initial_max     = 40,
 A_y             = 40,
 geno_length     = 2,
 g               = 9.81,
 remove          = 0.5,
 add_proportion  = 0.5,
 mutation_size   = 0.0000007,
 generations     = 100,
 population_size = 3000,
 maximise        = FALSE,
 location        = 0,
 minkowski_p     = 100
 ),
 store_data = FALSE,
 initial = initial_rand_uni_3d
);
#> Error in initialize(...): object 'initial_rand_uni_3d' not found

## ------------------------------------------------
## Method `ga$add_operators`
## ------------------------------------------------

brach$add_operators(
  operators = list(
    mutation_cauchy,
    crossover_basic_3d
  )
)
#> Error in eval(expr, envir, enclos): object 'brach' not found

brach$add_operators(
   mutation_cauchy
)
#> Error in eval(expr, envir, enclos): object 'brach' not found

## ------------------------------------------------
## Method `ga$add_metric`
## ------------------------------------------------

self$add_metric(name = "average_fitness", value = sum(fitnesses) / length(fitnesses));
#> Error in eval(expr, envir, enclos): object 'self' not found

## ------------------------------------------------
## Method `ga$add_dependents`
## ------------------------------------------------

brach$add_dependents(
  dependents = list(
    fitness_function_single = utility_fitness_brachistochrone_3d,
    fitness_function = utility_fitness_population_3d
  )
)
#> Error in eval(expr, envir, enclos): object 'brach' not found

## ------------------------------------------------
## Method `ga$add_dependent`
## ------------------------------------------------

brach$add_dependents(
  dependents = list(
    fitness_function_single = utility_fitness_brachistochrone_3d
  )
)
#> Error in eval(expr, envir, enclos): object 'brach' not found

## ------------------------------------------------
## Method `ga$next_operator`
## ------------------------------------------------

brach$next_operator()
#> Error in eval(expr, envir, enclos): object 'brach' not found

## ------------------------------------------------
## Method `ga$next_iteration`
## ------------------------------------------------

brach$next_iteration()
#> Error in eval(expr, envir, enclos): object 'brach' not found

## ------------------------------------------------
## Method `ga$generate_server`
## ------------------------------------------------

brach$generate_server()
#> Error in eval(expr, envir, enclos): object 'brach' not found

## ------------------------------------------------
## Method `ga$add_output_graph`
## ------------------------------------------------

brach$add_output_graph(
 list(
   xlab = "horizontal distance",
   ylab = "vertical distance",
   FUN = manhattan_plot
 )
)
#> Error in eval(expr, envir, enclos): object 'brach' not found

## ------------------------------------------------
## Method `ga$add_output_graphs`
## ------------------------------------------------

brach$add_output_graphs(
 list(
   list(
     xlab = "horizontal distance",
     ylab = "vertical distance",
     FUN = manhattan_plot
   ),
   list(
     xlab = "horizontal distance",
     ylab = "vertical distance",
     FUN = manhattan_plot
   )
 )
)
#> Error in eval(expr, envir, enclos): object 'brach' not found